# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"database/sql"
)

type Order struct {
	ID          int32        `json:"id"`
	CustomerID  int32        `json:"customer_id"`
	Status      string       `json:"status"`
	TotalAmount string       `json:"total_amount"`
	CreatedAt   sql.NullTime `json:"created_at"`
	UpdatedAt   sql.NullTime `json:"updated_at"`
}